% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggden-package.R
\docType{package}
\name{ggden-package}
\alias{ggden}
\alias{ggden-package}
\title{ggden: DEN House-Style Theme for ggplot2}
\description{
A ggplot2 theme and colour palettes following the visual identity of Dewan Ekonomi Nasional (DEN). Provides theme_den(), discrete and continuous colour/fill scales, and helper functions for consistent, publication-ready charts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/den-econ/ggden}
}

}
\author{
\strong{Maintainer}: Dewan Ekonomi Nasional \email{info@dewanekonomi.go.id}

Authors:
\itemize{
  \item Dewan Ekonomi Nasional \email{info@dewanekonomi.go.id}
}

}
\keyword{internal}
