% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_den}
\alias{scale_color_den}
\alias{scale_colour_den}
\title{DEN discrete colour scale}
\usage{
scale_color_den(...)

scale_colour_den(...)
}
\arguments{
\item{...}{Additional arguments passed to
\code{\link[ggplot2]{discrete_scale}}.}
}
\value{
A ggplot2 scale object.
}
\description{
Maps discrete aesthetics to the DEN palette.
}
\examples{
library(ggplot2)
ggplot(iris, aes(Sepal.Length, Sepal.Width, colour = Species)) +
  geom_point() +
  scale_color_den() +
  theme_den()
}
