% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_den}
\alias{scale_fill_den}
\title{DEN discrete fill scale}
\usage{
scale_fill_den(...)
}
\arguments{
\item{...}{Additional arguments passed to
\code{\link[ggplot2]{discrete_scale}}.}
}
\value{
A ggplot2 scale object.
}
\description{
Maps discrete fill aesthetics to the DEN palette.
}
\examples{
library(ggplot2)
ggplot(mpg, aes(class, fill = class)) +
  geom_bar() +
  scale_fill_den() +
  theme_den()
}
